done v 6 add little fixes

+++ src/main/java/com/work/exchangeRate/model/Rate.java
@@ -2,18 +2,19 @@ package com.work.exchangeRate.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Rate {
 
-    private HashMap<String, Double> rates;
+    private HashMap<String, BigDecimal> rates;
 
-    public HashMap<String, Double> getRates() {
+    public HashMap<String, BigDecimal> getRates() {
         return rates;
     }
 
-    public void setRates(HashMap<String, Double> rates) {
+    public void setRates(HashMap<String, BigDecimal> rates) {
         this.rates = rates;
     }
 }
